refactor: remove re-exports from modals index file
Eliminated re-exports from the modals index file to enhance module clarity and structure. Updated all related imports to use specific modal paths, improving maintainability and reducing ambiguity.
