--- apps/nextjs/src/components/features/modals/index.ts
@@ -1,8 +1,0 @@
-export * from './delete-task-section-modal';
-export * from './editor-link-modal';
-export * from './file-viewer-modal';
-export * from './invite-modal';
-export * from './modals';
-export * from './project-detail-modal';
-export * from './share-project-modal';
-export * from './share-workspace-modal';
