+++ apps/nextjs/src/components/features/navigation/projects/teammates.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { useInviteModal } from '@/components/features/modals';
+import { useInviteModal } from '@/components/features/modals/invite-modal/use-invite-modal';
 import { PADDING_X } from '@/components/features/navigation/navigation';
 import { TeammateAvatar } from '@/components/features/teammate-avatar';
 import { Wrap, WrapItem } from '@/components/ui/wrap';
