+++ apps/nextjs/src/components/features/popovers/popover-project-menu/share.tsx
@@ -1,6 +1,6 @@
 import type React from 'react';
 import { useCallback } from 'react';
-import { useShareProjectModal } from '@/components/features/modals';
+import { useShareProjectModal } from '@/components/features/modals/share-project-modal/use-share-project-modal';
 import { Menu } from '@/components/ui/menu';
 
 type Props = {
