+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/attachment/attachment.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { useFileViewerModal } from '@/components/features/modals';
+import { useFileViewerModal } from '@/components/features/modals/file-viewer-modal/use-file-viewer-modal';
 import { ThumbnailAttachment } from '@/components/features/thumbnail-attachment';
 import { Wrap, WrapItem } from '@/components/ui/wrap';
 import { useToaster } from '@/hooks/use-toaster';
