+++ apps/nextjs/src/components/features/task-detail/task-detail-body/form/feed-list/feed-list-item/content/content-attachment/content-attachment.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { useFileViewerModal } from '@/components/features/modals';
+import { useFileViewerModal } from '@/components/features/modals/file-viewer-modal/use-file-viewer-modal';
 import { FileTypeCode } from '@/graphql/enums';
 import {
   useTaskFile,
