+++ apps/nextjs/src/components/features/tasks/tasks-board/tasks-board-list-section/header/more-action/menu-list/menu-list.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { useDeleteTaskSectionModal } from '@/components/features/modals';
+import { useDeleteTaskSectionModal } from '@/components/features/modals/delete-task-section-modal/use-delete-task-section-modal';
 import {
   useHasTasksByTaskSectionId,
   useTasksTaskSectionCommand,
