+++ apps/nextjs/src/components/pages/projects/components/header/more-action-icon-button/menu-list/edit-project-details/edit-project-details.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { useProjectDetailModal } from '@/components/features/modals';
+import { useProjectDetailModal } from '@/components/features/modals/project-detail-modal/use-project-detail-modal';
 import { Icon } from '@/components/ui/icon';
 import { Menu } from '@/components/ui/menu';
