+++ apps/nextjs/src/components/pages/projects/components/header/project-detail-icon-button/project-detail-icon-button.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { useProjectDetailModal } from '@/components/features/modals';
+import { useProjectDetailModal } from '@/components/features/modals/project-detail-modal/use-project-detail-modal';
 import { Icon } from '@/components/ui/icon';
 import { IconButton, type IconButtonProps } from '@/components/ui/icon-button';
