+++ apps/nextjs/src/components/pages/workspaces/components/overview/members-section/add-member-list-item.tsx
@@ -1,5 +1,5 @@
 import { memo, useCallback } from 'react';
-import { useInviteModal } from '@/components/features/modals';
+import { useInviteModal } from '@/components/features/modals/invite-modal/use-invite-modal';
 import { TeammateAvatar } from '@/components/features/teammate-avatar';
 import { Flex } from '@/components/ui/flex';
 import { Text } from '@/components/ui/text';
