+++ apps/nextjs/src/components/pages/workspaces/components/overview/overview-content/project-roles-section/project-role-add-member/project-role-add-member.tsx
@@ -1,5 +1,5 @@
 import { useCallback } from 'react';
-import { useShareProjectModal } from '@/components/features/modals';
+import { useShareProjectModal } from '@/components/features/modals/share-project-modal/use-share-project-modal';
 import { TeammateAvatar } from '@/components/features/teammate-avatar';
 import { Flex } from '@/components/ui/flex';
 import { Text } from '@/components/ui/text';
